refactor: ProductInfoSummary 타입을 전역 types 폴더로 이동
