+++ src/data/api.ts
@@ -3,7 +3,7 @@ import { END_POINTS } from "./endPoints";
 import type { GiftThemeType } from "@/types/GiftThemeType";
 import type { TrendingGiftsType } from "@/types/TrendingGiftsType";
 import type { User } from "@/types/User";
-import type { ProductInfoSummary } from "@/pages/OrderPage/components/ProductInfo/ProductInfo";
+import type { ProductInfoSummary } from "@/types/ProductInfoSummary";
 
 export const apiClient = axios.create({
   baseURL: import.meta.env.VITE_API_BASE_URL,
