+++ src/types/ProductInfoSummary.ts
@@ -1,0 +1,7 @@
+export interface ProductInfoSummary {
+  id: number;
+  name: string;
+  brandName: string;
+  price: number;
+  imageURL: string;
+}
